extract VM trait to it's own crate
