+++ test_vm/tests/power_scenario_tests.rs
@@ -16,13 +16,15 @@ use fvm_shared::econ::TokenAmount;
 use fvm_shared::sector::{RegisteredPoStProof, RegisteredSealProof};
 use fvm_shared::METHOD_SEND;
 use num_traits::Zero;
-use test_vm::expects::Expect;
-use test_vm::trace::ExpectInvocation;
 use test_vm::util::{
-    apply_ok, assert_invariants, create_accounts, create_miner, expect_invariants,
-    invariant_failure_patterns, miner_dline_info, serialize_ok,
+    assert_invariants, create_accounts, create_miner, expect_invariants,
+    invariant_failure_patterns, miner_dline_info,
 };
-use test_vm::{TestVM, FIRST_TEST_USER_ADDR, TEST_FAUCET_ADDR, VM};
+use test_vm::{TestVM, FIRST_TEST_USER_ADDR, TEST_FAUCET_ADDR};
+use vm_api::expects::Expect;
+use vm_api::trace::ExpectInvocation;
+use vm_api::util::{apply_ok, serialize_ok};
+use vm_api::VM;
 
 #[test]
 fn power_create_miner() {
