+++ test_vm/tests/terminate_test.rs
@@ -22,15 +22,17 @@ use fvm_shared::piece::PaddedPieceSize;
 use fvm_shared::sector::{RegisteredSealProof, StoragePower};
 use num_traits::cast::FromPrimitive;
 use std::ops::Neg;
-use test_vm::expects::Expect;
-use test_vm::trace::ExpectInvocation;
 use test_vm::util::{
     advance_by_deadline_to_epoch, advance_by_deadline_to_epoch_while_proving,
-    advance_to_proving_deadline, apply_ok, create_accounts, create_miner, expect_invariants,
-    get_state, invariant_failure_patterns, make_bitfield, market_publish_deal, miner_balance,
-    submit_windowed_post, verifreg_add_verifier, DynBlockstore,
+    advance_to_proving_deadline, create_accounts, create_miner, expect_invariants,
+    invariant_failure_patterns, make_bitfield, market_publish_deal, miner_balance,
+    submit_windowed_post, verifreg_add_verifier,
 };
-use test_vm::{TestVM, VM};
+use test_vm::TestVM;
+use vm_api::expects::Expect;
+use vm_api::trace::ExpectInvocation;
+use vm_api::util::{apply_ok, get_state, DynBlockstore};
+use vm_api::VM;
 
 #[test]
 fn terminate_sectors() {
