+++ test_vm/tests/verifreg_remove_datacap_test.rs
@@ -26,13 +26,12 @@ use fil_actors_runtime::{
 };
 use fvm_ipld_encoding::ipld_block::IpldBlock;
 use fvm_shared::address::Address;
-use test_vm::expects::Expect;
-use test_vm::trace::ExpectInvocation;
-use test_vm::util::{
-    apply_code, apply_ok, assert_invariants, create_accounts, get_state, verifreg_add_verifier,
-    DynBlockstore,
-};
-use test_vm::{TestVM, TEST_VERIFREG_ROOT_ADDR, VM};
+use test_vm::util::{assert_invariants, create_accounts, verifreg_add_verifier};
+use test_vm::{TestVM, TEST_VERIFREG_ROOT_ADDR};
+use vm_api::expects::Expect;
+use vm_api::trace::ExpectInvocation;
+use vm_api::util::{apply_code, apply_ok, get_state, DynBlockstore};
+use vm_api::VM;
 
 #[test]
 fn remove_datacap_simple_successful_path() {
