+++ vm_api/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "test_vm"
-description = "Reference vm for integration testing builtin actors"
-version = "12.0.0"
+name = "vm_api"
+description = "Abstract virtual machine interface used for testing FVM native wasm actors"
+version = "1.0.0"
 license = "MIT OR Apache-2.0"
 authors = ["Protocol Labs", "Filecoin Core Devs"]
 edition = "2021"
@@ -53,13 +53,4 @@ regex = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
 libsecp256k1 = { workspace = true }
-fil_actors_evm_shared = { workspace = true }
-vm_api = { workspace = true }
 
-[dev-dependencies]
-cid = { workspace = true }
-multihash = { workspace = true }
-test-case = { workspace = true }
-ethers = { workspace = true }
-hex = { workspace = true }
-hex-literal = { workspace = true }
