+++ vm_api/src/error.rs
@@ -1,0 +1,28 @@
+use std::{error::Error, fmt};
+
+#[derive(Debug)]
+pub struct VMError {
+    msg: String,
+}
+
+impl fmt::Display for VMError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.msg)
+    }
+}
+
+impl Error for VMError {
+    fn description(&self) -> &str {
+        &self.msg
+    }
+}
+
+impl From<fvm_ipld_hamt::Error> for VMError {
+    fn from(h_err: fvm_ipld_hamt::Error) -> Self {
+        vm_err(h_err.to_string().as_str())
+    }
+}
+
+pub fn vm_err(msg: &str) -> VMError {
+    VMError { msg: msg.to_string() }
+}
