+++ vm_api/src/expects.rs
