+++ vm_api/src/trace.rs
