+++ vm_api/src/util/blockstore.rs
